docs(wip): migrating-from-v1-to-v2
