+++ services/Website/website/i18n/en.json
@@ -1,0 +1,79 @@
+{
+  "_comment": "This file is auto-generated by write-translations.js",
+  "localized-strings": {
+    "next": "Next",
+    "previous": "Previous",
+    "tagline": "为跑团而生的即时通讯应用",
+    "docs": {
+      "actor/actor1": {
+        "title": "创建属于自己的人物卡!"
+      },
+      "actor/actor2": {
+        "title": "从一个简单的人物卡开始"
+      },
+      "actor/tags/baseattr": {
+        "title": "<BaseAttr />"
+      },
+      "actor/tags/baseinfo": {
+        "title": "<BaseInfo />"
+      },
+      "actor/tags/col": {
+        "title": "<Col />"
+      },
+      "actor/tags/row": {
+        "title": "<Row />"
+      },
+      "actor/tags/tabs": {
+        "title": "<Tabs />"
+      },
+      "actor/tags/template": {
+        "title": "<Template />"
+      },
+      "config": {
+        "title": "配置"
+      },
+      "develop": {
+        "title": "开发"
+      },
+      "getstart": {
+        "title": "起步",
+        "sidebar_label": "从这开始"
+      },
+      "doc2": {
+        "title": "document number 2"
+      },
+      "doc3": {
+        "title": "This is document number 3"
+      },
+      "doc5": {
+        "title": "Fifth Document"
+      },
+      "introduce": {
+        "title": "介绍"
+      },
+      "package-core": {
+        "title": "核心包"
+      }
+    },
+    "links": {
+      "文档": "文档",
+      "开发": "开发",
+      "帮助": "帮助",
+      "博客": "博客",
+      "Github": "Github",
+      "关于作者": "关于作者"
+    },
+    "categories": {
+      "开始": "开始",
+      "人物": "人物",
+      "标签": "标签",
+      "开发": "开发",
+      "包": "包"
+    }
+  },
+  "pages-strings": {
+    "Help Translate|recruit community translators for your project": "Help Translate",
+    "Edit this Doc|recruitment message asking to edit the doc source": "Edit",
+    "Translate this Doc|recruitment message asking to translate the docs": "Translate"
+  }
+}
