+++ services/Website/website/package.json
@@ -1,18 +1,17 @@
 {
   "scripts": {
-    "examples": "docusaurus-examples",
-    "start": "docusaurus-start",
-    "build": "docusaurus-build",
-    "publish-gh-pages": "docusaurus-publish",
-    "write-translations": "docusaurus-write-translations",
-    "version": "docusaurus-version",
-    "rename-version": "docusaurus-rename-version"
+    "start": "docusaurus start",
+    "build": "docusaurus build",
+    "swizzle": "docusaurus swizzle",
+    "deploy": "docusaurus deploy"
   },
-  "devDependencies": {
-    "docusaurus": "^1.11.1"
-  },
+  "devDependencies": {},
   "dependencies": {
+    "@docusaurus/core": "^2.0.0-alpha.50",
+    "@docusaurus/preset-classic": "^2.0.0-alpha.50",
     "highlight.js": "^9.18.1",
-    "lz-string": "^1.4.4"
+    "lz-string": "^1.4.4",
+    "react": "^16.8.4",
+    "react-dom": "^16.8.4"
   }
 }
