+++ services/Website/website/src/pages/index.js
@@ -1,0 +1,214 @@
+import React from 'react';
+import useDocusaurusContext from '@docusaurus/useDocusaurusContext';
+import Layout from '@theme/Layout';
+
+const MarkdownBlock = (props) => (
+  <div>{props.children}</div>
+); /* Used to read markdown */
+const Container = (props) => <div>{props.children}</div>;
+const GridBlock = (props) => <div>{props.children}</div>;
+
+class HomeSplash extends React.Component {
+  render() {
+    const { siteConfig, language = '' } = this.props;
+    const { baseUrl, docsUrl } = siteConfig;
+    const docsPart = `${docsUrl ? `${docsUrl}/` : ''}`;
+    const langPart = `${language ? `${language}/` : ''}`;
+    const docUrl = (doc) => `${baseUrl}${docsPart}${langPart}${doc}`;
+
+    const SplashContainer = (props) => (
+      <div className="homeContainer">
+        <div className="homeSplashFade">
+          <div className="wrapper homeWrapper">{props.children}</div>
+        </div>
+      </div>
+    );
+
+    const Logo = (props) => (
+      <div className="projectLogo">
+        <img src={props.img_src} alt="Project Logo" />
+      </div>
+    );
+
+    const ProjectTitle = () => (
+      <h2 className="projectTitle">
+        {siteConfig.title}
+        <small>{siteConfig.tagline}</small>
+      </h2>
+    );
+
+    const PromoSection = (props) => (
+      <div className="section promoSection">
+        <div className="promoRow">
+          <div className="pluginRowBlock">{props.children}</div>
+        </div>
+      </div>
+    );
+
+    const Button = (props) => (
+      <div className="pluginWrapper buttonWrapper">
+        <a className="button" href={props.href} target={props.target}>
+          {props.children}
+        </a>
+      </div>
+    );
+
+    return (
+      <SplashContainer>
+        <Logo img_src={`${baseUrl}img/undraw_chatting_2yvo.svg`} />
+        <div className="inner">
+          <ProjectTitle siteConfig={siteConfig} />
+          <PromoSection>
+            <Button href="https://trpg.moonrailgun.com">网页版</Button>
+            {/* <Button href="/download">下载桌面版</Button> */}
+            <Button href="https://trpg.moonrailgun.com/portal/deploy">
+              下载移动版
+            </Button>
+          </PromoSection>
+        </div>
+      </SplashContainer>
+    );
+  }
+}
+
+function Index() {
+  const context = useDocusaurusContext();
+  const { siteConfig } = context;
+  const { baseUrl, language } = siteConfig;
+
+  const Block = (props) => (
+    <div
+      padding={['bottom', 'top']}
+      id={props.id}
+      background={props.background}
+    >
+      <GridBlock
+        align="center"
+        contents={props.children}
+        layout={props.layout}
+      />
+    </div>
+  );
+
+  const FeatureCallout = () => (
+    <div
+      className="productShowcaseSection paddingBottom"
+      style={{ textAlign: 'center' }}
+    >
+      <h2>更多特性</h2>
+      <MarkdownBlock>
+        `易用`,`简单`,`便捷`. 一直在试图为用户带来最好的跑团体验
+      </MarkdownBlock>
+    </div>
+  );
+
+  const TryOut = () => (
+    <Block id="try">
+      {[
+        {
+          title: '高度可定制',
+          content:
+            '创建自己的脚本，人物。 搭建自己专属的团。你能想到的。我们都会为你提供！',
+          image: `${baseUrl}img/undraw_code_review.svg`,
+          imageAlign: 'left',
+        },
+      ]}
+    </Block>
+  );
+
+  const Description = () => (
+    <Block background="dark">
+      {[
+        {
+          title: '便捷',
+          content:
+            '走到哪，聊到哪。不论是车上，床上，食堂。都能保持与朋友的沟通不被中断',
+          image: `${baseUrl}img/undraw_note_list.svg`,
+          imageAlign: 'right',
+        },
+      ]}
+    </Block>
+  );
+
+  const LearnHow = () => (
+    <Block background="light">
+      {[
+        {
+          title: '即时沟通',
+          content: '不论你在哪，都能和天南海北的伙伴一起享受跑团的乐趣.',
+          image: `${baseUrl}img/undraw_youtube_tutorial.svg`,
+          imageAlign: 'right',
+        },
+      ]}
+    </Block>
+  );
+
+  // 特性
+  const Features = () => (
+    <Block layout="fourColumn">
+      {[
+        {
+          title: '现代化',
+          content:
+            '代码由现代流行的React打造，不管是开发与使用都能为你带来现代化的用户与开发体验',
+          image: `${baseUrl}img/undraw_react.svg`,
+          imageAlign: 'top',
+        },
+        {
+          title: '高度活跃',
+          content:
+            '在过去的几年中，长久保持高活跃度的开发进行与产品迭代。争取为用户提供最好的体验',
+          image: `${baseUrl}img/undraw_operating_system.svg`,
+          imageAlign: 'top',
+        },
+      ]}
+    </Block>
+  );
+
+  const Showcase = () => {
+    if ((siteConfig.users || []).length === 0) {
+      return null;
+    }
+
+    const showcase = siteConfig.users
+      .filter((user) => user.pinned)
+      .map((user) => (
+        <a href={user.infoLink} key={user.infoLink}>
+          <img src={user.image} alt={user.caption} title={user.caption} />
+        </a>
+      ));
+
+    const pageUrl = (page) => baseUrl + (language ? `${language}/` : '') + page;
+
+    return (
+      <div className="productShowcaseSection paddingBottom">
+        <h2>谁在使用?</h2>
+        <p>这个项目正在被这些人使用</p>
+        <div className="logos">{showcase}</div>
+        <div className="more-users">
+          <a className="button" href={pageUrl('users.html')}>
+            更多 {siteConfig.title} 用户
+          </a>
+        </div>
+      </div>
+    );
+  };
+
+  return (
+    <Layout>
+      <div>
+        {/* <HomeSplash siteConfig={siteConfig} language={language} /> */}
+        <div className="mainContainer">
+          {/* <Features />
+          <FeatureCallout />
+          <LearnHow />
+          <TryOut />
+          <Description />
+          <Showcase /> */}
+        </div>
+      </div>
+    </Layout>
+  );
+}
+
+export default Index;
