+++ services/Website/website/src/pages/users.js
@@ -1,15 +1,6 @@
-/**
- * Copyright (c) 2017-present, Facebook, Inc.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- */
+import React from 'react';
 
-const React = require('react');
-
-const CompLibrary = require('../../core/CompLibrary.js');
-
-const Container = CompLibrary.Container;
+const Container = (props) => <div>{props.children}</div>;
 
 class Users extends React.Component {
   render() {
@@ -45,4 +36,4 @@ class Users extends React.Component {
   }
 }
 
-module.exports = Users;
+export default Users;
